Minimal change to fix the problem with pollingGetActivity$ that "completes" when we get a 403 from the server.
The problem was that the checkConnection() call threw an exception when
the connectionStatus was ExpiredToken. This resulted in the outer catch
in the pollingGetActivity() function getting called which returned an
empty Observable that "completed" the activity$ stream.

A temporary fix for this problem is to not throw from the checkConnection
function - Also, we must not make AJAX calls while the connection status
is not online.